Add user organizations (#7454)
Connects: https://github.com/department-of-veterans-affairs/caseflow/issues/7455
Resolves: https://github.com/department-of-veterans-affairs/caseflow/issues/7427

### Description
This PR creates a new Organization-User join table to manage what users are part of which organization. It also creates a UI at `/organizations/<org url>/users` to manage the users in an organization. Any user within the organization can add or remove other users from this page.

Next steps: we need to change our organization class to return the members of the organization from this new join table instead of using VACOLS.

### Testing Plan
1. Log in as the org queue user
1. Go to `/organizations/translation/users`
1. Add and remove users to your hearts content

![organizationusers](https://user-images.githubusercontent.com/3885236/47184912-f7bf5000-d2f9-11e8-9f53-dd0484be9699.gif)