persist appeals when searching by vbms id (#7598)
Resolves https://github.com/department-of-veterans-affairs/caseflow/issues/7579

### Description
If multiple appeals are passed with `nil` values for id to our serializers, the serializer drops all but one appeal (it treats `nil` as a unique id even though it's not). As a result, if more than one appeal without a LegacyAppeal row in our DB was found in case search, all but one would be dropped by the serializer and not shown in case search results.

We need to save appeals before passing them to the serializer.