Only show available_actions for active tasks (#7649)
Limited in scope to only GenericTasks and ColocatedTasks for now since we expect those to be the most frequent occurrences. GenericTasks will be assigned through organizations which should not have available actions when there are active child tasks (this PR fixes this situation), and ColocatedTasks should not be actionable once they have been returned to the assigning attorney even though colocated staffers will still be able to view the appeals to which they are connected.