explain how to fix security audit failures (#7751)
### Description
We've seen a couple times where Brakeman fails CI on a security audit of the installed gems. Adding an explanation of how to address when this comes up so people don't get stuck.

In addition to the audit findings (which list a CVE and the gem in violation), the rake task will also print the following:
```
Failed. Security vulnerabilities were found. Find the dependency in Gemfile.lock, then specify a safe version of the dependency in the Gemfile (preferred) or snooze the CVE in security.rake for a week.
See https://github.com/department-of-veterans-affairs/caseflow/pull/7639 for an example.
```

### Acceptance Criteria
N/A

### Testing Plan
N/A