Tries alternative claim number if ssn does not work in VBMS request (#7755)
Resolves #4567

### Description
In some situations the file number in VACOLS (the `bfcorlid` column) is incorrect. It will be the SSN where it should be a claim number. When hitting the VBMS endpoint to get a list of files, we use the wrong number, and get an error with text: "File Number does not exist within the system." To solve this problem we can make a call to BGS to find the correct file number and then call VBMS with that instead.