Add Assigned to for AMA/RAMP appeals (#7783)
Resolves https://github.com/department-of-veterans-affairs/caseflow/issues/7360

### Description
Populate the CSS ID of whoever has the most recent task for an appeal in the Assigned to column of the Case list for AMA appeals. If case is assigned to the current user, show `Assigned to you`. If case has not distributed yet, show `Case storage`, and show `Post decision` if the root task is marked as completed.

<img width="960" alt="screenshot 2018-11-08 16 32 39" src="https://user-images.githubusercontent.com/2928395/48228782-2c13b280-e374-11e8-9dc0-25f0e3e9ad59.png">

@lowellrex — please let me know any way I can clean up any of the `.rb` here — including if there's more idiomatic ways to write the tests and if there's a way to clean up the logic in `appeal.rb`.

/cc @laurjpeterson