Change display name of VLJ support team (colocated) (#7901)
During a meeting with several Board teams yesterday, a Board representative informed us that the VLJ support management is now referred to as the "VLJ Support Staff". This PR changes their display name throughout the application ("assign to team" task actions, co-located team queue, and one place in the copy of the case details page).