Include named subclasses in list of assignable organizations (#7933)
Noticed this bug when preparing my local environment for a demo with the mail team this morning. The list of organizations returned by the [where statement in `assignable()`](https://github.com/department-of-veterans-affairs/caseflow/blob/952ab605bf2aa6a1f7dbc7366de06784c24781a8/app/models/organization.rb#L7) excludes organizations that are subclasses of the main Organization class unless they are explicitly included in that list. This was a result of a decision made [when we initially added](https://github.com/department-of-veterans-affairs/caseflow/pull/7183#pullrequestreview-160016295) the ability to list assignable organizations.

This change fixes the immediate problem and aims to be more flexible to the addition of new organizations going forward.