Fix React bug with state and props out of sync (#7993)
connects #7772 

### Description
The refactor in 5270977f8e4cbbea0c2f4f1a5ff5474fcd21b96c introduced a regression due to how React components are rendered. When a list of child components is rendered, and then one is removed, all the child components bump up the stack. That meant that issues might inherit the state of the component just before them, causing them to appear as eligibility errors when they are not.

The solution is to calculate eligibility errors at every `render` instead of assuming that an issue and a component will stay together forever.

### Acceptance Criteria
- [x] Code compiles correctly