Set timeout to 5 min (new_documents and document_count) and catch error properly (#8168)
Resolves #7945

### Description
The call to new_documents and document_count can last up to 4 minutes. From New Relic

![image](https://user-images.githubusercontent.com/3811786/49707849-23c0c880-fbfb-11e8-9f31-20f01673b682.png)

![image](https://user-images.githubusercontent.com/3811786/49707855-27ece600-fbfb-11e8-802d-0d5d5e9a4af7.png)

Nginx has default timeout of 60 seconds for its connection to the server side. If the back end server does not respond in 60 seconds, Nginx tears down its connection to the server, and further the connection from the client.
We can increase it by setting `proxy_read_timeout`. As an example from the devops repo:
https://github.com/department-of-veterans-affairs/devops/blob/04e08585ffea68e4477c12cb75d4a8b2756c78ee/ansible/roles/revproxy-configure/templates/nginx_revproxy.conf#L210

Increase timeout on the client for the above calls and catch error properly.

![image](https://user-images.githubusercontent.com/3811786/49707881-4652e180-fbfb-11e8-975c-586d9ecc2bf5.png)

Debugging further shows that a lot of requests are in the stalled state

```
Stalled/Blocking
Time the request spent waiting before it could be sent. This time is inclusive of any time spent in proxy negotiation. Additionally, this time will include when the browser is waiting for an already established connection to become available for re-use, obeying Chrome's maximum six TCP connection per origin rule.
```

![image](https://user-images.githubusercontent.com/3811786/49746692-bd29c200-fc6f-11e8-96e6-7107e303b73c.png)