Clear Redis namespaced keys before and after tests
**Why**:
`Rails.cache.clear` runs `flushdb`, which only clears the current Redis DB, not
all keys across all namespaces. This can lead to issues if the test DB got in a
bad state. For example, if a spec enables a particular feature toggle but
forgets to turn it off, it will be turned on for other tests that might depend
on it being turned off. Running `Rails.cache.clear` will not reset the Redis
namespace where the toggle state is stored.

**How**:
For each namespace we know about, delete the keys in Redis. Normally, we would
pass in the namespace option to `Rails.cache.clear`, but it doesn’t work in
versions of Rails lower than 5.2: https://github.com/rails/rails/issues/33216

Until we update to Rails 5.2, we can use `scan_each`. H/T @saneshark for this:
https://github.com/department-of-veterans-affairs/vets-api/pull/421/files