Convert decass notes to ASCII before saving to DB (#8349)
Resolves #8128

**Why**:
The VACOLS.DECASS.DEATCOM DB column is limited to 350 characters. If an attorney
submits a note with 350 characters, where some are multi-byte characters, such
as curly quotes or accented characters, the VACOLS Oracle DB will treat the
string as having more than 350 characters, which results in an exception due to
the DB constraint.

**How**:
- On the backend, only use the first 350 characters, then convert them all to
ASCII using the `stringex` gem which has a handy `to_ascii` method.

On the frontend, the Notes field was already limited to 350 characters.

The truncation and ASCII conversion happens in the `QueueMapper` class, which
was refactored using the following principles:

- Use a class instead of a module to avoid passing arguments around
- Give the class a single public method and only test that method
- Move the data transformation to individual short methods with clear names
- Keep all methods under 10 lines of code
- Order methods by level of abstraction for better readability

I noticed that other VACOLS columns have length limits, so we might need to
convert those to ASCII as well. Or perhaps it might make sense to have a generic
ASCII transformation that gets applied to all VACOLS string columns.