Show correct actions if judge assigns a QR task to an attorney (#8407)
When the Quality Review team returns a case to a judge we create a `JudgeAssignTask`. If a `JudgeAssignTask` has a parent task of the type `QualityReviewTask` then we display an option to "Mark task complete" in order to return the case to the Quality Review team after the judge has addressed the concerns of the Quality Review team. However, if the judge assigns one of these cases to an attorney on their team (in the parlance of the application, creates a child `AttorneyTask`) then we change the type of the judge task from `JudgeAssignTask` to `JudgeReviewTask` [after the attorney returns the case to the judge](https://github.com/department-of-veterans-affairs/caseflow/blob/53dc4c07c63e195130d6ad9973a9515871bf1941/app/models/tasks/judge_task.rb#L23) (completes the `AttorneyTask`). Because we change the type of task, we no longer display the "Mark task complete" option. This PR changes that to correctly display the "Mark task complete" option in those cases.

This is just a temporary bugfix that should be addressed in a more thorough way by #8107.