Move condition to decide if we show/hide task action dropdown to backend (#8441)
This PR moves logic onto the backend now that we have a robust model for determining available actions for each task. When this PR was created we were still using the [combination of individual, attorney, and organization tasks](https://github.com/department-of-veterans-affairs/caseflow/pull/7275/files#diff-5b5c7473c1568400a8104eccd0693108R241) in order to determine whether to show the task actions dropdown. Now that we are using the `Task. available_actions()` and the `actionableTasksForAppeal()` selector, we can move a bunch of this code to the backend.

Before | After
--- | ---
![image](https://user-images.githubusercontent.com/32683958/50454549-308d2e00-0916-11e9-9648-d2dd19946389.png) | ![image](https://user-images.githubusercontent.com/32683958/50454533-22d7a880-0916-11e9-9e9e-302f0b52a0d5.png)