Remove unused response status (#8443)
In the process of writing a tech spec to unify all of the various `app/models/queue/` models we have I noticed that we return a status code that is never used by the front-end. This PR removes the unused response status in order to prepare for a refactor of the `TasksController` that will probably result from the aforementioned tech spec.