Only return tasks from LegacyWorkQueue methods (#8444)
In the process of writing a tech spec to unify all of the various `app/models/queue/` models we have I noticed that the `LegacyWorkQueue` methods return a list of two elements but we only ever use the first (tasks). This PR refactors the `LegacyWorkQueue` methods to only return the first element in that last.