Move automatic task assignment logic from task to organization (#8489)
While working on [automatically routing mail tasks to different teams](https://github.com/department-of-veterans-affairs/caseflow/pull/8222) based on the state of the appeal, I identified the way we currently handle automatic assignment of tasks from organizations to members of those organizations as an area for potential improvement.

This PR changes the way we determine which members of the `Colocated` team we automatically assign `ColocatedTask`s to. Prior to this PR, the task was responsible for determining which `Colocated` employee to assign the task to (by way of the `RoundRobinAssigner` concern). With this PR, the responsibility for determining which `Colocated` employee to assign the task to is shifted onto the `Colocated` organization (by way of the `ColocatedTaskAssigneeFactory` class).

The two major benefits of this change are:
1. The application logic maps more neatly onto the business logic. We would like to automatically assign all tasks that are assigned to the `Colocated` team to a member of that team without regard for the nature of that task (though we are still able to do that by [passing a `task_type` argument](https://github.com/department-of-veterans-affairs/caseflow/pull/8489/files#diff-e9b0bdeeb6d95d19520c1006758e2672R24) to the method which returns us the next assignee).
1. We can now just assign tasks to an organization and allow that assignee to determine whether the task should create a child task. This will save us from [replicating code](https://github.com/department-of-veterans-affairs/caseflow/pull/8222/files#diff-738d69bc1fa41b93704d5dcc4d805639R67) whenever we want to reproduce the automatic task assignment functionality for a single organization.

Additionally, by moving this automatic assignment logic to the `Organization` model, we could use its access control features to enable admins to determine how tasks should be routed within that organization (maybe to exclude a new team member from the automatic task assignment rotation for their first week, or to specifically route all "new evidence" tasks to one member of the team).

### Next steps (later PRs)
* Use this pattern to [replace `RoundRobinAssigner` entirely](https://github.com/department-of-veterans-affairs/caseflow/blob/3009822a3cbde29f7dcaee6da46751637a393658/app/models/concerns/round_robin_assigner.rb) in [`JudgeTask`](https://github.com/department-of-veterans-affairs/caseflow/blob/3009822a3cbde29f7dcaee6da46751637a393658/app/models/tasks/judge_task.rb) and [`BvaDispatchTask`](https://github.com/department-of-veterans-affairs/caseflow/blob/3009822a3cbde29f7dcaee6da46751637a393658/app/models/tasks/bva_dispatch_task.rb)
* Use the `Colocated` organization's automatic task assignment logic instead of the [existing `ColocatedTask` task assignment logic](https://github.com/department-of-veterans-affairs/caseflow/blob/3009822a3cbde29f7dcaee6da46751637a393658/app/models/tasks/colocated_task.rb#L22)