Add test for task instructions shape change (#8538)
While working on #7906 I tried to move the `modify_params` method to a `before_create` hook, but that didn't work because I guess ActiveRecord enforces type parity before the `before_create` hook runs (alternatively we could override `Task`'s instructions setter method to do this, but I don't think we use that pattern yet in our code and I'm wary of introducing it). I think this test is still valuable, so I'm inclined to leave it in.