Creates distribution, evidence window, and informal hearing presentation tasks for direct review and evidence submission appeals  (#8333)
Resolves https://github.com/department-of-veterans-affairs/caseflow/issues/8334
Related to https://github.com/department-of-veterans-affairs/caseflow/issues/7916 but does not complete it, will be continued in future PR

### Description
If we have ama auto case distribution on, create Distribution, IHP, and evidence submission window tasks as necessary. This PR just creates the tasks, later PRs will introduce the logic to expire/complete them. 

### Testing Plan
1. unit tests pass