issue.eligibleForSocOptIn is undefined for RequestIssues on edit (#8812)
connects #7337 

### Description
Fix subtle JS bug where `undefined !== false`.

This is because the same code gets applied to `ContestedIssue` and `RequestIssue` when both creating and editing an Intake. On edit, the `RequestIssue` has an undefined value for `eligibleForSocOptIn` and was triggering an ineligible warning message for the issue, even though it was eligible.