Fix document count missing in case details spec
**Details**:
The `AppealDocumentCount` React component does not attempt to fetch
the document count if the appeal's `file_type` is "Paper". In this
particular spec, the file type was "Paper", and so the document count
wouldn't show up.

In order to make this a non-paper file type, we can use either the
`:docs_in_vbms` or `:docs_in_vva` trait for the `case` factory. This is
because the file type gets set based on VACOLS attributes when the
`AppealsController` is called.

The flow is as follows:
1. `AppealsController#index` calls `get_appeals_for_file_number`
2. `LegacyAppeal.fetch_appeals_by_file_number` is called
3. `AppealRepository.appeals_by_vbms_id` is called
4. `AppealRepository.build_appeal` is called
4. `AppealRepository.set_vacols_values` is called
5. `file_type` is set to `AppealRepository.folder_type_from`
Unless the properties `tivbms` or `tisubj2` are set to specific values,
the `file_type` will be set to "Paper".