hide tasks that are on hold in the judge view of attorney caseloads (#8957)
Resolves https://github.com/department-of-veterans-affairs/caseflow/issues/8318
/cc @laurjpeterson 

### Description
This PR hides cases that are on hold from judge view of attorney caseloads. This PR follows @amprokop's proposed approach in https://github.com/department-of-veterans-affairs/caseflow/issues/8318.

When judges are assigning cases to their attorneys, they need to see an accurate number of cases that are on their attorney's plates.