Fix default payee code (#8935)
connects #8933 

*Description*

The `end_product_establishments.payee_code` column was getting null values. This PR adds a db constraint to prevent that.

The dilemma is that we are using `DecisionReview.payee_code` to set the value of new `EndProductEstablishment` records, but we are also using it to populate the client UI form. We want the UI form to default to `nil` so that the user is prompted to select a value, but that means that `nil` values may somehow slip into the db.

What we do not have is a reproducible test for how nil values get into the db. The hope is that by adding the db constraint we can discover the code path that is leading to that state.