make board grants only show specific business line (#9050)
Connects #8299

Task has a decision_review, which has all the request issues regardless of type. But appeals can have multiple tasks assigned to each business line. This filters out the request issues that are of other benefit types, which prevents the decision issues of the other business types from displaying.