Don't update AttorneyCaseReview if not found (#9066)
**Why**: We came across a bug where someone attempted to update the
document ID for a case that only existed in VACOLS and had not gone
through the Caseflow checkout yet. This meant that a corresponding entry
in the AttorneyCaseReview table didn't exist. Previously, we assumed
there would always be a corresponding entry. When one did not exist,
the code tried to call `update!` on `nil`.

**How**
- If a corresponding entry in AttorneyCaseReview is not found, don't
attempt to update or create a new entry.
- Update the list of work product codes that correspond to draft
decisions by using the DECASS_WORK_PRODUCT_TYPES constant.

Fixes this Sentry issue:
https://sentry.ds.va.gov/department-of-veterans-affairs/caseflow/issues/3521/