Remove unused `judge_css_id` method from User
As far as I can tell, this method is not being used.
It was first introduced in
https://github.com/department-of-veterans-affairs/caseflow/pull/6161 to
support
https://github.com/department-of-veterans-affairs/caseflow/issues/5174,
but I don’t see in that PR how that is being used by the front end.
From what I can tell, the user dropdown that judges use to assign cases
displays the full name of users, not their CSS ID.

By removing this method, we eliminate a bunch of unnecessary DB calls.