Close tracking tasks when we close the root task (#9173)
Connects #7905. This PR adds code to close `TrackVeteranTask`s related to an appeal when that appeal is outcoded (and creates a method on `RootTask` that we can use to attach other events to if we'd like). This change limits the set of appeals that will show up in a VSO's "All cases" view now because we will be closing the tracking tasks. Additionally, it reduces the population of tasks that we will need to cancel when an appellant's representation changes (this code doesn't exist, but will be completed by part of #7905).