Prevent saving case with missing required fields
**Why**: During attorney checkout, if a required field is missing, we
want the frontend to catch that to avoid a backend request. As for the
document ID validation, we allow a server request because the backend
validation includes helpful error messages which aren't implemented on
the frontend, and we didn't think it was worth it to duplicate the
logic.

Resolves #9063

Acceptance Criteria:
When submitting a decision, whether a draft decision or an OMO,
submitting the form without all required fields filled in does not
make a backend request. Error messages appear inline to higlight which
fields are missing.