Add closed_at column to tasks table (#9180)
Connects #8701. Adds the `closed_at` column in preparation to replace the `completed_at` column so that the `completed` and `closed` task statuses can set the same timestamp column. Removing the `completed_at` column will follow in a later PR that will be merged at least 1 day after this change is live to allow enough time for the migration.