get decision date based on decision issues (#9111)
Connects #8031

Tasks don't have decision date, so find it based off of the decision issue. Decision issue dates are saved as UTC and we convert to local time in the browser, which throws off the date. Modified the date format to be able to get a date from UTC time.