automatically create translation tasks for veterans in Puerto Rico or Philippines (#9191)
Resolves #8226

### Description
Establishes new `TranslationTask` task type, and automatically assigns a `TranslationTask` to the Translation organization on successful intake if the veteran's address is in Puerto Rico or Philippines.

The code uses the [`VADotGovService`](https://github.com/department-of-veterans-affairs/caseflow/blob/master/app/services/external_api/va_dot_gov_service.rb) to validate the veteran's address. If the service is unavailable, it fails silently and falls back to the veteran's address. It does not retry.