Use !empty? or .any? instead of .count > 0 (#9234)
**Why**: When checking if an array contains elements, `!empty` (or
`size > 0`) is faster than `count > 0`. `count` should rarely be used in
a Rails application.

When checking if an ActiveRecord::Relation contains results, `.any?` is
preferred because it runs a `SELECT 1 AS one LIMIT 1` SQL query as
opposed to a `COUNT`, which can be very slow on a large table. In all
of the instances changed in this PR, we don't care about the actual
count of elements, so `count` is not the right choice.

Recommended reading:
https://www.speedshop.co/2019/01/10/three-activerecord-mistakes.html