Add check for nil (#9300)
https://github.com/department-of-veterans-affairs/appeals-support/issues/4083

VBMS has two ids for each document. A series_id and a version_id. Multiple document versions can belong to the same series, but each version gets its own version. When VBMS made this change we created a method to copy over metadata (like comments) from old versions of the documents to the new versions since we didn't want people to loose their comments/tags/etc. However, it seems like a lot of documents now have a series_id of null. So we began associating null series_id documents with each other and copying over the metadata.

This PR is a short term fix, which explicitly checks for null and doesn't copy things over if it is null. The next step is to dig in further and figure out why we're getting so many nulls.