VA.gov will need the docket type to be in camelCase (#9329)
So my ticket was wrong. The change is pretty straightforward. Here's why I think it's needed. TL;DR: The mismatch of snake_case for values and camelCase for keys finally bit us.

After MVP, we'll be adding a `eta` hash to the `docket` hash. I anticipated that it would look like:

```json
"eta": {
  "direct_review": "2020-01-01",
  "new_evidence": "2025-01-01",
  "hearing": "2025-01-01"
}
```

This enables us to look up `docket.type` in `docket.eta` to find the ETA of the currently selected docket, as well as to use all of the other estimates to provide alternative ETA if the Veteran wants to switch dockets.

Only it won't look like that, because we have our serializer set to turn all keys into camelCase. So it comes out like:

```json
"eta": {
  "directReview": "2020-01-01",
  "newEvidence": "2025-01-01",
  "hearing": "2025-01-01"
}
```

Bleargh. Now we end up needing to do a whole bunch of mapping between snake_case and camelCase depending on whether what we're looking at came from a key or a value.

We can solve it by just changing `docket.type` to use the same values as we'll use for the keys of `docket.eta`. It's compounding the inconsistency, but it's certainly a lot less messy on the front end, and not too weird on the Caseflow side, I think.