Add indexes to Tasks table
**Why**: Sentry has been reporting statement timeout errors on queries
for the Tasks table, and New Relic confirms that those are slow queries.

Indexes have been added for the most common slow queries on the Tasks
table.

I've also set Postgres to automatically analyze the table after adding
an index. The `strong_migrations` gem provides this functionality.
From the Postgres documentation:

```
ANALYZE collects statistics about the contents of tables in the
database, and stores the results in the pg_statistic system catalog.
Subsequently, the query planner uses these statistics to help determine
the most efficient execution plans for queries.
```

Resolves #9233