Add fasterer gem to catch performance issues (#9308)
**Why**:
In Ruby and in Rails, there are multiple ways of achieving of
the same result. However, they are not always equal in terms of
performance, and choosing the wrong one can have a significant impact.

For example, we recently had daily DB statement timeouts between January
22 and February 6 that caused the AWS RDS CPU utilization to increase by
about 5-10x the normal rate. This was caused by using `.count > 0` to
check if an ActiveRecord query contained any results. This query was on
the Documents table, which has been quickly growing in size. `COUNT`
queries in Postgres are known to be slow, and get slower as the table
grows.

If all we want to know is if the query returned any results at all, that
means we don't care about the actual count, and therefore we don't need
to perform a `COUNT` query. Instead, we can use
`ActiveRecord::Relation#any?` which performs a much faster SQL query
that uses `SELECT 1 AS one` with a `LIMIT` of 1.

The [fasterer](https://github.com/DamirSvrtan/fasterer) gem was inspired
by the [fast-ruby](https://github.com/JuanitoFatas/fast-ruby) repo,
which in turn was inspired by a great
[talk on Ruby performance](https://www.youtube.com/watch?v=fGFM_UrSp70)
by Erik Michaels-Ober.

It analyzes the code for slow idioms and lets you know which version is
faster. This is a great way to learn something new and to level up the
whole team. The gem doesn't currently track all of the examples in the
fast-ruby repo, but it does track some of the most common mistakes. I
plan on contributing to the gem to add more checks.

The 4 most common slow idioms we had were:

- `select`ing from an array and then calling `.first` or `.last` instead
of `detect`
- Using `Hash#merge!` instead of `Hash#[]=` when only merging one
element
- Sorting using the idiom `sort { |a, b| a.foo <=> b.foo }` instead of
`sort_by(&:foo)`. Relatedly, using `sort_by` and then `.first` or
`.last` instead of `min_by` and `max_by` respectively.
- Passing in a block argument and using `block.call` instead of just
calling `yield`.

In most cases, the faster code also had the additional advantage of making the
code easier to read and understand.