Fix flakey queue specs
For the attorney checkout specs, I tried a bunch of things, including
adding unique class names and IDs to the checkboxes and radio buttons,
but the only thing that seemed to work is to call `page.all` an extra
time, possibly combined with changing the order of clicks. For example,
instead of clicking "Current findings" and then one of the radio buttons
beneath it, we click "Current findings", then "Nexus option", then the
radio buttons under each of those checkboxes.

Resolves #8680