Treat EPE w/o a synced status as active (#9391)
connects #9376 

### Description
Fixes a race condition. When an `EndProductEstablishment` has been created but not yet established upstream, treat it as active for the purposes of identifying duplicate request issues on future intakes.

**Why**

Because EPE is performed asynchronously, there is an unspecified period of time between creating the EPE and it receiving a synced status after establishing upstream End Product(s). During that period of time, we don't want another request issue created that would eventually be a duplicate.