Fix for showing draft dispositions through API (#9410)
The Board received a report of a VA.gov user being able to see a draft decision on the Issues tab. In addition to recording the draft disposition, VACOLS is marking the issue as having a close date, bypassing the existing check (unclear whether this has always been the case or is a regression). This adds a check of the status of the appeal itself to ensure that we do not show Board dispositions (allowed, denied, remanded) of an issue while it is active at the Board.