Exclude appeals that require(d) IHPs for the purposes of calculating … (#9447)
Noticed during testing that `DirectReviewDocket.new.time_until_due_of_oldest_appeal` was returning 0. Turns out that the receipt dates of the oldest ten appeals are: [Thu, 05 Apr 2018, Tue, 10 Apr 2018, Wed, 16 May 2018, Fri, 01 Jun 2018, Thu, 07 Jun 2018, Sat, 30 Jun 2018, Sun, 01 Jul 2018, Tue, 31 Jul 2018, Sun, 05 Aug 2018, Mon, 06 Aug 2018]

This is because they're blocked by IHPs, which the Board has yet to negotiate deadlines for.

The obvious fix is just to exclude non-ready appeals. This PR does that.

But when an IHP is completed for one of these old appeals, it will become ready, and the next time someone requests a distribution, the docket will suddenly appear dramatically older than it in fact is, overdistributing direct review appeals.

So this PR further proposes excluding appeals that have ever had an IHP task from the calculation.

I'll be talking this through with LRP in a bit. But wanted to get the PR in early.