Fix CaseDocket specs
**Why**:
Date fields in VACOLS are in the format `Sun, 04 Dec 2016`. However,
dates like `1.year.ago` are in the format
`Thu, 15 Feb 2018 17:08:50 UTC +00:00`. The
`nonpriority_decisions_per_year` method was causing an Oracle error
because of the `1.year.ago` query, which was later fixed by adding
`.to_date`. The reason why the specs in  `case_docket_spec.rb` were
passing originally is because there was a `before` block that called
`Timecop.freeze(Time.utc(2015, 1, 1, 12, 0, 0))`. This causes all `ago`
dates to be in the format that Oracle expects.

I also added an extra `case` with `aod` in the spec in order to test the
`AOD = 0` condition.