Avoid calling BGS when finding a Veteran record (#9515)
https://rpm.newrelic.com/accounts/1788458/applications/80608544/transactions#id=5b22436f6e74726f6c6c65722f7461736b732f72656164795f666f725f68656172696e675f7363686564756c65222c22225d

### Description
Hearings schedule pages are spending a substantial percentage of their time calling BGS. This is because of a recently added log statement that makes a call to BGS even when our veteran record is in our database. This adds a check around the log statement, so we only do it when necessary, and adds a test to ensure we don't erroneously make calls to BGS in this way again.