Allow generic document uploads to VBMS via IDT
**Why**: Paul Saindon has requested the ability to upload any type of
document to VBMS via an IDT endpoint.

**How**:
- Create a new `vbms_uploaded_documents` DB table with various
timestamps used by the `Asyncable` module to track the upload process,
similar to the `decision_documents` table (which might be refactored
later to use this new table).
- Create a new endpoint for uploading documents at:
`idt/api/v1/appeals/:appeal_id/upload_document`. This endpoint requires
the following parameters: `appeal_id` (the appeal's UUID), `file`: a
Base64-encoded file, and `document_type`, a string description of the
document.
- The controller calls a Service Object that validates that the appeal
exists, that a file is present, and that the `document_type` is
registered in the Caseflow Commons Document Types:
https://github.com/department-of-veterans-affairs/caseflow-commons/blob/master/app/models/caseflow/document_types.rb
If everything is valid, the file is uploaded to VBMS via a background
job.

Resolves #8641

Test Plan:
Work with Paul to test doc uploads to VBMS in UAT